+++ psml_web.py
@@ -13,7 +13,7 @@ session_opts = {
     'sessioni.auto':True
 }
 
-    
+
 @route('/script/<filepath:path>')
 def server_static(filepath):
     return static_file(filepath, root=os.path.dirname(__file__))
@@ -134,8 +134,8 @@ def change():
 
 app=default_app()
 app=SessionMiddleware(app, session_opts)
-if __name__=="__main__":
-    run_server()
-
 def run_server():
     run(app=app,host="127.0.0.1")
+
+if __name__=="__main__":
+    run_server()